feat(jupyterlab:docker): use java headless JRE instead of JDK
Users are not expected to develop Java based tools (or rebuild the
NeuroML stack) on the platform, so we don't need to install the JDK.
Similarly, the GUI elements of jNeuroML etc. won't work on the platform,
so we can use the headless JRE.

This should reduce the number of deps pulled in and save us some space.