feat(repository-page): add information tooltip
This adds an icon next to the "Overview" text on the repository page.
The icon has a tooltip to tell users that the repository page always
shows the latest version/contents of a repository, and that to see
previous versions/contents, they can visit the repository at it's
source.

Fixes #342