fix(workspace-page): propagate workspace page errors to ErrorBoundary
If we check that `workspace` is `null` before rendering, when the
`WorkspaceService` calls fail and `workspace` is `null`, we end up with
a blank page because nothing is rendered.

One way of handling this is to check each service call. A better way,
however, is to let the React Components error and then let the
ErrorBoundary catch these errors.

So, here, we no longer check for `workspace` to not be `null` before
rendering.