feat(workspace-page): set most recently used app as first option
First, instead of hard-coding apps here, we use the OSBApplications
dictionary. This way, if a new app is used, only the OSBApplications
dictionary needs to be updated.

Next, using this dictionary, we construct the text to be shown in the
options, using the most recently used app as the first option if it
exists.

Finally, to open the app, because we know the option text format, we can
quickly check the string and open the necessary app.

Fixes #370