feat(add-workspace): use consistent dialogs for workspace creation
While the "create repository" dialog is written as a separate
component, the "create workspace" dialog was using the `OSBDialog`
component. Here, the buttons were being passed as part of the `OSBDialog`
children ending up in `DialogContent` instead of as `DialogActions`. As
a result, they were in the inner container and required scrolling
instead in the outer `OSBDialog` container as buttons.

So, I've just refactored the code to make the actions all fall into
`OSBDialog` actions.

Fixes #405