fix(repository-page): ensure that WorkspaceEditor is rendered fresh
Here we must use `&& showWorkspaceEditor` so that the `WorkspaceEditor`
component is rendered afresh each time. If we don't do this, the states
in the `WorkspaceEditor` component like `workspaceForm` are initialised
only once with the value of `workspace` prop---at the initial
render---and then do not track the value of the `workpace` prop.

An alternative would be to use a `useEffect` in `WorkspaceEditor` to
update the value of the state variable from the props each time.