feat(user-page): move dialog bits to user-page from user-editor
This ensures that the actions of the OSBDialog are used---so they're in
the outer component and do not get hidden and require scrolling. This
also ensures that the page doesn't need to reload etc. when the edit
profile dialog is cancelled without any changes (this was flashing up an
error too).