fix(biomodels): replace '+' in file names with spaces
Fixes #943

The Biomodels API returns files with '+' in their names instead of
spaces. While this is a good idea, the issue is that the name of the
downloaded files are different from what the web interface shows. This
is unexpected.

So, we manually rename the files back to the spaced versions after
downloading them into a temporary directory. This assumes that:

- no file uses '+' in its file name: they're all only there as
  replacements for spaces

We must download to a temporary directory because otherwise we'll have
to figure out a way of tracking what files were downloaded and only
rename them. Otherwise, using the wildcard in `rename` would rename all
files in the folder, even ones that weren't downloaded in this session.
While it is unlikely that users will use '+' in their own file names, if
they do, we don't want to tinker with that.